+++ src/taylor.cpp
@@ -235,10 +235,10 @@ llvm::Value *taylor_c_diff_numparam_codegen(llvm_state &s, const param &, llvm::
 namespace
 {
 
-// Add a function for computing the continuous output
+// Add a function for computing the dense output
 // via polynomial evaluation.
 template <typename T>
-void taylor_add_c_out_function(llvm_state &s, std::uint32_t n_eq, std::uint32_t order, std::uint32_t batch_size)
+void taylor_add_d_out_function(llvm_state &s, std::uint32_t n_eq, std::uint32_t order, std::uint32_t batch_size)
 {
     assert(n_eq > 0u);
     assert(order > 0u);
@@ -257,11 +257,11 @@ void taylor_add_c_out_function(llvm_state &s, std::uint32_t n_eq, std::uint32_t 
     auto *ft = llvm::FunctionType::get(builder.getVoidTy(), fargs, false);
     assert(ft != nullptr);
     // Now create the function.
-    auto *f = llvm::Function::Create(ft, llvm::Function::ExternalLinkage, "c_out_f", &s.module());
+    auto *f = llvm::Function::Create(ft, llvm::Function::ExternalLinkage, "d_out_f", &s.module());
     // LCOV_EXCL_START
     if (f == nullptr) {
         throw std::invalid_argument(
-            "Unable to create a function for the continuous output in an adaptive Taylor integrator");
+            "Unable to create a function for the dense output in an adaptive Taylor integrator");
     }
     // LCOV_EXCL_STOP
 
@@ -1061,7 +1061,7 @@ void taylor_adaptive_impl<T>::finalise_ctor_impl(U sys, std::vector<T> state, T 
 
     // Temporarily disable optimisations in s, so that
     // we don't optimise twice when adding the step
-    // and then the c_out.
+    // and then the d_out.
     std::optional<opt_disabler> od(m_llvm);
 
     // Add the stepper function.
@@ -1069,8 +1069,8 @@ void taylor_adaptive_impl<T>::finalise_ctor_impl(U sys, std::vector<T> state, T 
         = taylor_add_adaptive_step<T>(m_llvm, "step", std::move(sys), tol, 1, high_accuracy, compact_mode);
 
     // Add the function for the computation of
-    // the continuous output.
-    taylor_add_c_out_function<T>(m_llvm, m_dim, m_order, 1);
+    // the dense output.
+    taylor_add_d_out_function<T>(m_llvm, m_dim, m_order, 1);
 
     // Restore the original optimisation level in s.
     od.reset();
@@ -1084,8 +1084,8 @@ void taylor_adaptive_impl<T>::finalise_ctor_impl(U sys, std::vector<T> state, T 
     // Fetch the stepper.
     m_step_f = reinterpret_cast<step_f_t>(m_llvm.jit_lookup("step"));
 
-    // Fetch the function to compute the continuous output.
-    m_c_out_f = reinterpret_cast<c_out_f_t>(m_llvm.jit_lookup("c_out_f"));
+    // Fetch the function to compute the dense output.
+    m_d_out_f = reinterpret_cast<d_out_f_t>(m_llvm.jit_lookup("d_out_f"));
 
     // Setup the vector for the Taylor coefficients.
     // LCOV_EXCL_START
@@ -1098,18 +1098,18 @@ void taylor_adaptive_impl<T>::finalise_ctor_impl(U sys, std::vector<T> state, T 
 
     m_tc.resize(m_state.size() * (m_order + 1u));
 
-    // Setup the vector for the continuous output.
-    m_c_out.resize(m_state.size());
+    // Setup the vector for the dense output.
+    m_d_out.resize(m_state.size());
 }
 
 template <typename T>
 taylor_adaptive_impl<T>::taylor_adaptive_impl(const taylor_adaptive_impl &other)
     // NOTE: make a manual copy of all members, apart from the function pointers.
     : m_state(other.m_state), m_time(other.m_time), m_llvm(other.m_llvm), m_dim(other.m_dim), m_dc(other.m_dc),
-      m_order(other.m_order), m_pars(other.m_pars), m_tc(other.m_tc), m_last_h(other.m_last_h), m_c_out(other.m_c_out)
+      m_order(other.m_order), m_pars(other.m_pars), m_tc(other.m_tc), m_last_h(other.m_last_h), m_d_out(other.m_d_out)
 {
     m_step_f = reinterpret_cast<step_f_t>(m_llvm.jit_lookup("step"));
-    m_c_out_f = reinterpret_cast<c_out_f_t>(m_llvm.jit_lookup("c_out_f"));
+    m_d_out_f = reinterpret_cast<d_out_f_t>(m_llvm.jit_lookup("d_out_f"));
 }
 
 template <typename T>
@@ -1293,17 +1293,17 @@ std::uint32_t taylor_adaptive_impl<T>::get_dim() const
 }
 
 template <typename T>
-const std::vector<T> &taylor_adaptive_impl<T>::update_c_output(T time)
+const std::vector<T> &taylor_adaptive_impl<T>::update_d_output(T time)
 {
     // NOTE: "time" needs to be translated
-    // because m_c_out_f expects a time coordinate
+    // because m_d_out_f expects a time coordinate
     // with respect to the starting time t0 of
     // the *previous* timestep. Thus, we need to compute:
     const auto h = time - (m_time - m_last_h);
 
-    m_c_out_f(m_c_out.data(), m_tc.data(), &h);
+    m_d_out_f(m_d_out.data(), m_tc.data(), &h);
 
-    return m_c_out;
+    return m_d_out;
 }
 
 // Explicit instantiation of the implementation classes/functions.
@@ -1425,7 +1425,7 @@ void taylor_adaptive_batch_impl<T>::finalise_ctor_impl(U sys, std::vector<T> sta
 
     // Temporarily disable optimisations in s, so that
     // we don't optimise twice when adding the step
-    // and then the c_out.
+    // and then the d_out.
     std::optional<opt_disabler> od(m_llvm);
 
     // Add the stepper function.
@@ -1433,8 +1433,8 @@ void taylor_adaptive_batch_impl<T>::finalise_ctor_impl(U sys, std::vector<T> sta
         = taylor_add_adaptive_step<T>(m_llvm, "step", std::move(sys), tol, m_batch_size, high_accuracy, compact_mode);
 
     // Add the function for the computation of
-    // the continuous output.
-    taylor_add_c_out_function<T>(m_llvm, m_dim, m_order, m_batch_size);
+    // the dense output.
+    taylor_add_d_out_function<T>(m_llvm, m_dim, m_order, m_batch_size);
 
     // Restore the original optimisation level in s.
     od.reset();
@@ -1448,8 +1448,8 @@ void taylor_adaptive_batch_impl<T>::finalise_ctor_impl(U sys, std::vector<T> sta
     // Fetch the stepper.
     m_step_f = reinterpret_cast<step_f_t>(m_llvm.jit_lookup("step"));
 
-    // Fetch the function to compute the continuous output.
-    m_c_out_f = reinterpret_cast<c_out_f_t>(m_llvm.jit_lookup("c_out_f"));
+    // Fetch the function to compute the dense output.
+    m_d_out_f = reinterpret_cast<d_out_f_t>(m_llvm.jit_lookup("d_out_f"));
 
     // Setup the vector for the Taylor coefficients.
     // LCOV_EXCL_START
@@ -1467,10 +1467,10 @@ void taylor_adaptive_batch_impl<T>::finalise_ctor_impl(U sys, std::vector<T> sta
     // Setup m_last_h.
     m_last_h.resize(boost::numeric_cast<decltype(m_last_h.size())>(batch_size));
 
-    // Setup the vector for the continuous output.
+    // Setup the vector for the dense output.
     // NOTE: the size of m_state.size() already takes
     // into account the batch size.
-    m_c_out.resize(m_state.size());
+    m_d_out.resize(m_state.size());
 
     // Prepare the temp vectors.
     m_pinf.resize(m_batch_size, std::numeric_limits<T>::infinity());
@@ -1487,7 +1487,7 @@ void taylor_adaptive_batch_impl<T>::finalise_ctor_impl(U sys, std::vector<T> sta
     m_cur_max_delta_ts.resize(m_batch_size);
     m_pfor_ts.resize(m_batch_size);
 
-    m_c_out_time.resize(boost::numeric_cast<decltype(m_step_res.size())>(m_batch_size));
+    m_d_out_time.resize(boost::numeric_cast<decltype(m_step_res.size())>(m_batch_size));
 }
 
 template <typename T>
@@ -1495,13 +1495,13 @@ taylor_adaptive_batch_impl<T>::taylor_adaptive_batch_impl(const taylor_adaptive_
     // NOTE: make a manual copy of all members, apart from the function pointers.
     : m_batch_size(other.m_batch_size), m_state(other.m_state), m_time(other.m_time), m_llvm(other.m_llvm),
       m_dim(other.m_dim), m_dc(other.m_dc), m_order(other.m_order), m_pars(other.m_pars), m_tc(other.m_tc),
-      m_last_h(other.m_last_h), m_c_out(other.m_c_out), m_pinf(other.m_pinf), m_minf(other.m_minf),
+      m_last_h(other.m_last_h), m_d_out(other.m_d_out), m_pinf(other.m_pinf), m_minf(other.m_minf),
       m_delta_ts(other.m_delta_ts), m_step_res(other.m_step_res), m_prop_res(other.m_prop_res),
       m_ts_count(other.m_ts_count), m_min_abs_h(other.m_min_abs_h), m_max_abs_h(other.m_max_abs_h),
-      m_cur_max_delta_ts(other.m_cur_max_delta_ts), m_pfor_ts(other.m_pfor_ts), m_c_out_time(other.m_c_out_time)
+      m_cur_max_delta_ts(other.m_cur_max_delta_ts), m_pfor_ts(other.m_pfor_ts), m_d_out_time(other.m_d_out_time)
 {
     m_step_f = reinterpret_cast<step_f_t>(m_llvm.jit_lookup("step"));
-    m_c_out_f = reinterpret_cast<c_out_f_t>(m_llvm.jit_lookup("c_out_f"));
+    m_d_out_f = reinterpret_cast<d_out_f_t>(m_llvm.jit_lookup("d_out_f"));
 }
 
 template <typename T>
@@ -1791,29 +1791,29 @@ std::uint32_t taylor_adaptive_batch_impl<T>::get_dim() const
 }
 
 template <typename T>
-const std::vector<T> &taylor_adaptive_batch_impl<T>::update_c_output(const std::vector<T> &time)
+const std::vector<T> &taylor_adaptive_batch_impl<T>::update_d_output(const std::vector<T> &time)
 {
     // Check the dimensionality of time.
     if (time.size() != m_batch_size) {
         using namespace fmt::literals;
 
         throw std::invalid_argument(
-            "Invalid number of time coordinates specified for the continuous output in a Taylor integrator in batch "
+            "Invalid number of time coordinates specified for the dense output in a Taylor integrator in batch "
             "mode: the batch size is {}, but the number of time coordinates is {}"_format(m_batch_size, time.size()));
     }
 
     // NOTE: "time" needs to be translated
-    // because m_c_out_f expects a time coordinate
+    // because m_d_out_f expects a time coordinate
     // with respect to the starting time t0 of
     // the *previous* timestep. Thus, we need to compute:
     // time - (m_time - m_last_h);
     for (std::uint32_t i = 0; i < m_batch_size; ++i) {
-        m_c_out_time[i] = time[i] - (m_time[i] - m_last_h[i]);
+        m_d_out_time[i] = time[i] - (m_time[i] - m_last_h[i]);
     }
 
-    m_c_out_f(m_c_out.data(), m_tc.data(), m_c_out_time.data());
+    m_d_out_f(m_d_out.data(), m_tc.data(), m_d_out_time.data());
 
-    return m_c_out;
+    return m_d_out;
 }
 
 // Explicit instantiation of the batch implementation classes.