+++ test/taylor_adaptive.cpp
@@ -12,6 +12,7 @@
 #include <sstream>
 #include <stdexcept>
 #include <string>
+#include <tuple>
 #include <utility>
 #include <vector>
 
@@ -30,6 +31,18 @@
 
 using namespace heyoka;
 using namespace heyoka_test;
+
+template <typename Out, typename P, typename T>
+auto &horner_eval(Out &ret, const P &p, int order, const T &eval)
+{
+    ret = xt::view(p, xt::all(), order);
+
+    for (--order; order >= 0; --order) {
+        ret = xt::view(p, xt::all(), order) + ret * eval;
+    }
+
+    return ret;
+}
 
 TEST_CASE("streaming op")
 {
@@ -240,7 +253,7 @@ TEST_CASE("last h")
     }
 }
 
-TEST_CASE("continuous output")
+TEST_CASE("dense output")
 {
     auto [x, v] = make_vars("x", "v");
 
@@ -254,22 +267,22 @@ TEST_CASE("continuous output")
                                                   kw::compact_mode = cm,
                                                   kw::opt_level = opt_level};
 
-                auto c_out = xt::adapt(ta.get_c_output());
+                auto d_out = xt::adapt(ta.get_d_output());
 
                 // Take a first step.
                 ta.step(true);
 
-                // The c_out at t = 0 must be the same
+                // The d_out at t = 0 must be the same
                 // as the IC.
-                ta.update_c_output(0);
-                REQUIRE(c_out[0] == approximately(0.05, 100.));
-                REQUIRE(c_out[1] == approximately(0.025, 100.));
+                ta.update_d_output(0);
+                REQUIRE(d_out[0] == approximately(0.05, 100.));
+                REQUIRE(d_out[1] == approximately(0.025, 100.));
 
-                // The c_out at the end of the timestep must be
+                // The d_out at the end of the timestep must be
                 // equal to the current state.
-                ta.update_c_output(ta.get_time());
-                REQUIRE(c_out[0] == approximately(ta.get_state()[0], 100.));
-                REQUIRE(c_out[1] == approximately(ta.get_state()[1], 100.));
+                ta.update_d_output(ta.get_time());
+                REQUIRE(d_out[0] == approximately(ta.get_state()[0], 100.));
+                REQUIRE(d_out[1] == approximately(ta.get_state()[1], 100.));
 
                 // Store the state at the end of the first step.
                 auto old_state1 = ta.get_state();
@@ -277,18 +290,18 @@ TEST_CASE("continuous output")
                 // Take a second step.
                 ta.step(true);
 
-                // The c_out at the beginning of the timestep
+                // The d_out at the beginning of the timestep
                 // must be equal to the state at the end of the
                 // previous timestep.
-                ta.update_c_output(ta.get_time() - ta.get_last_h());
-                REQUIRE(c_out[0] == approximately(old_state1[0], 100.));
-                REQUIRE(c_out[1] == approximately(old_state1[1], 100.));
+                ta.update_d_output(ta.get_time() - ta.get_last_h());
+                REQUIRE(d_out[0] == approximately(old_state1[0], 100.));
+                REQUIRE(d_out[1] == approximately(old_state1[1], 100.));
 
-                // The c_out at the end of the timestep must be
+                // The d_out at the end of the timestep must be
                 // equal to the current state.
-                ta.update_c_output(ta.get_time());
-                REQUIRE(c_out[0] == approximately(ta.get_state()[0], 100.));
-                REQUIRE(c_out[1] == approximately(ta.get_state()[1], 100.));
+                ta.update_d_output(ta.get_time());
+                REQUIRE(d_out[0] == approximately(ta.get_state()[0], 100.));
+                REQUIRE(d_out[1] == approximately(ta.get_state()[1], 100.));
 
                 // Store the state at the end of the second timestep.
                 auto old_state2 = ta.get_state();
@@ -296,18 +309,18 @@ TEST_CASE("continuous output")
                 // Take a third timestep.
                 ta.step(true);
 
-                // The c_out at the beginning of the timestep
+                // The d_out at the beginning of the timestep
                 // must be equal to the state at the end of the
                 // previous timestep.
-                ta.update_c_output(ta.get_time() - ta.get_last_h());
-                REQUIRE(c_out[0] == approximately(old_state2[0], 100.));
-                REQUIRE(c_out[1] == approximately(old_state2[1], 100.));
+                ta.update_d_output(ta.get_time() - ta.get_last_h());
+                REQUIRE(d_out[0] == approximately(old_state2[0], 100.));
+                REQUIRE(d_out[1] == approximately(old_state2[1], 100.));
 
-                // The c_out at the end of the timestep must be
+                // The d_out at the end of the timestep must be
                 // equal to the current state.
-                ta.update_c_output(ta.get_time());
-                REQUIRE(c_out[0] == approximately(ta.get_state()[0], 100.));
-                REQUIRE(c_out[1] == approximately(ta.get_state()[1], 100.));
+                ta.update_d_output(ta.get_time());
+                REQUIRE(d_out[0] == approximately(ta.get_state()[0], 100.));
+                REQUIRE(d_out[1] == approximately(ta.get_state()[1], 100.));
 
                 // Do it a few more times.
                 for (auto i = 0; i < 100; ++i) {
@@ -317,13 +330,13 @@ TEST_CASE("continuous output")
 
                     REQUIRE(oc == taylor_outcome::success);
 
-                    ta.update_c_output(ta.get_time() - ta.get_last_h());
-                    REQUIRE(c_out[0] == approximately(old_state2[0], 1000.));
-                    REQUIRE(c_out[1] == approximately(old_state2[1], 1000.));
+                    ta.update_d_output(ta.get_time() - ta.get_last_h());
+                    REQUIRE(d_out[0] == approximately(old_state2[0], 1000.));
+                    REQUIRE(d_out[1] == approximately(old_state2[1], 1000.));
 
-                    ta.update_c_output(ta.get_time());
-                    REQUIRE(c_out[0] == approximately(ta.get_state()[0], 1000.));
-                    REQUIRE(c_out[1] == approximately(ta.get_state()[1], 1000.));
+                    ta.update_d_output(ta.get_time());
+                    REQUIRE(d_out[0] == approximately(ta.get_state()[0], 1000.));
+                    REQUIRE(d_out[1] == approximately(ta.get_state()[1], 1000.));
                 }
             }
         }
@@ -351,17 +364,17 @@ TEST_CASE("continuous output")
 
                     auto sa = xt::adapt(ta.get_state_data(), {2u, batch_size});
                     auto isa = xt::adapt(init_state.data(), {2u, batch_size});
-                    auto coa = xt::adapt(ta.get_c_output(), {2u, batch_size});
+                    auto coa = xt::adapt(ta.get_d_output(), {2u, batch_size});
 
                     ta.step(true);
 
-                    ta.update_c_output(std::vector<double>(batch_size, 0.));
+                    ta.update_d_output(std::vector<double>(batch_size, 0.));
                     for (auto i = 0u; i < batch_size; ++i) {
                         REQUIRE(coa(0u, i) == approximately(isa(0u, i), 100.));
                         REQUIRE(coa(1u, i) == approximately(isa(1u, i), 100.));
                     }
 
-                    ta.update_c_output(ta.get_time());
+                    ta.update_d_output(ta.get_time());
                     for (auto i = 0u; i < batch_size; ++i) {
                         REQUIRE(coa(0u, i) == approximately(sa(0u, i), 100.));
                         REQUIRE(coa(1u, i) == approximately(sa(1u, i), 100.));
@@ -373,13 +386,13 @@ TEST_CASE("continuous output")
 
                     ta.step(true);
 
-                    ta.update_c_output(old_time1);
+                    ta.update_d_output(old_time1);
                     for (auto i = 0u; i < batch_size; ++i) {
                         REQUIRE(coa(0u, i) == approximately(ost1a(0u, i), 100.));
                         REQUIRE(coa(1u, i) == approximately(ost1a(1u, i), 100.));
                     }
 
-                    ta.update_c_output(ta.get_time());
+                    ta.update_d_output(ta.get_time());
                     for (auto i = 0u; i < batch_size; ++i) {
                         REQUIRE(coa(0u, i) == approximately(sa(0u, i), 100.));
                         REQUIRE(coa(1u, i) == approximately(sa(1u, i), 100.));
@@ -391,13 +404,13 @@ TEST_CASE("continuous output")
 
                     ta.step(true);
 
-                    ta.update_c_output(old_time2);
+                    ta.update_d_output(old_time2);
                     for (auto i = 0u; i < batch_size; ++i) {
                         REQUIRE(coa(0u, i) == approximately(ost2a(0u, i), 100.));
                         REQUIRE(coa(1u, i) == approximately(ost2a(1u, i), 100.));
                     }
 
-                    ta.update_c_output(ta.get_time());
+                    ta.update_d_output(ta.get_time());
                     for (auto i = 0u; i < batch_size; ++i) {
                         REQUIRE(coa(0u, i) == approximately(sa(0u, i), 100.));
                         REQUIRE(coa(1u, i) == approximately(sa(1u, i), 100.));
@@ -410,13 +423,13 @@ TEST_CASE("continuous output")
 
                         ta.step(true);
 
-                        ta.update_c_output(old_time);
+                        ta.update_d_output(old_time);
                         for (auto i = 0u; i < batch_size; ++i) {
                             REQUIRE(coa(0u, i) == approximately(osta(0u, i), 10000.));
                             REQUIRE(coa(1u, i) == approximately(osta(1u, i), 10000.));
                         }
 
-                        ta.update_c_output(ta.get_time());
+                        ta.update_d_output(ta.get_time());
                         for (auto i = 0u; i < batch_size; ++i) {
                             REQUIRE(coa(0u, i) == approximately(sa(0u, i), 10000.));
                             REQUIRE(coa(1u, i) == approximately(sa(1u, i), 10000.));
@@ -426,17 +439,175 @@ TEST_CASE("continuous output")
                     using Catch::Matchers::Message;
 
                     REQUIRE_THROWS_MATCHES(
-                        ta.update_c_output({}), std::invalid_argument,
-                        Message("Invalid number of time coordinates specified for the continuous output in a Taylor "
+                        ta.update_d_output({}), std::invalid_argument,
+                        Message("Invalid number of time coordinates specified for the dense output in a Taylor "
                                 "integrator in batch "
                                 "mode: the batch size is "
                                 + std::to_string(batch_size) + ", but the number of time coordinates is 0"));
                     REQUIRE_THROWS_MATCHES(
-                        ta.update_c_output(std::vector<double>(123u)), std::invalid_argument,
-                        Message("Invalid number of time coordinates specified for the continuous output in a Taylor "
+                        ta.update_d_output(std::vector<double>(123u)), std::invalid_argument,
+                        Message("Invalid number of time coordinates specified for the dense output in a Taylor "
                                 "integrator in batch "
                                 "mode: the batch size is "
                                 + std::to_string(batch_size) + ", but the number of time coordinates is 123"));
+                }
+            }
+        }
+    }
+}
+
+TEST_CASE("taylor tc basic")
+{
+    // Scalar test.
+    for (auto opt_level : {0u, 1u, 2u, 3u}) {
+        for (auto cm : {false, true}) {
+            for (auto ha : {false, true}) {
+                auto [x, v] = make_vars("x", "v");
+
+                auto ta = taylor_adaptive<double>{{prime(x) = v, prime(v) = -9.8 * sin(x)},
+                                                  {0.05, 0.025},
+                                                  kw::high_accuracy = ha,
+                                                  kw::compact_mode = cm,
+                                                  kw::opt_level = opt_level};
+
+                REQUIRE(ta.get_tc().size() == 2u * (ta.get_order() + 1u));
+
+                auto tca = xt::adapt(ta.get_tc().data(), {2u, ta.get_order() + 1u});
+
+                auto [oc, h] = ta.step(true);
+
+                auto ret = xt::eval(xt::zeros<double>({2}));
+
+                horner_eval(ret, tca, static_cast<int>(ta.get_order()), 0);
+                REQUIRE(ret[0] == approximately(0.05, 10.));
+                REQUIRE(ret[1] == approximately(0.025, 10.));
+
+                horner_eval(ret, tca, static_cast<int>(ta.get_order()), h);
+                REQUIRE(ret[0] == approximately(ta.get_state()[0], 10.));
+                REQUIRE(ret[1] == approximately(ta.get_state()[1], 10.));
+
+                auto old_state = ta.get_state();
+
+                std::tie(oc, h) = ta.step_backward(true);
+
+                horner_eval(ret, tca, static_cast<int>(ta.get_order()), 0);
+                REQUIRE(ret[0] == approximately(old_state[0], 10.));
+                REQUIRE(ret[1] == approximately(old_state[1], 10.));
+
+                horner_eval(ret, tca, static_cast<int>(ta.get_order()), h);
+                REQUIRE(ret[0] == approximately(ta.get_state()[0], 10.));
+                REQUIRE(ret[1] == approximately(ta.get_state()[1], 10.));
+
+                old_state = ta.get_state();
+
+                std::tie(oc, h) = ta.step(1e-3, true);
+
+                horner_eval(ret, tca, static_cast<int>(ta.get_order()), 0);
+                REQUIRE(ret[0] == approximately(old_state[0], 10.));
+                REQUIRE(ret[1] == approximately(old_state[1], 10.));
+
+                horner_eval(ret, tca, static_cast<int>(ta.get_order()), h);
+                REQUIRE(ret[0] == approximately(ta.get_state()[0], 10.));
+                REQUIRE(ret[1] == approximately(ta.get_state()[1], 10.));
+            }
+        }
+    }
+
+    // Batch test.
+    for (auto batch_size : {1u, 4u, 23u}) {
+        for (auto opt_level : {0u, 1u, 2u, 3u}) {
+            for (auto cm : {false, true}) {
+                for (auto ha : {false, true}) {
+                    auto [x, v] = make_vars("x", "v");
+
+                    std::vector<double> init_state;
+                    for (auto i = 0u; i < batch_size; ++i) {
+                        init_state.push_back(0.05 + i / 100.);
+                    }
+                    for (auto i = 0u; i < batch_size; ++i) {
+                        init_state.push_back(0.025 + i / 1000.);
+                    }
+
+                    auto ta = taylor_adaptive_batch<double>{{prime(x) = v, prime(v) = -9.8 * sin(x)},
+                                                            init_state,
+                                                            batch_size,
+                                                            kw::high_accuracy = ha,
+                                                            kw::compact_mode = cm,
+                                                            kw::opt_level = opt_level};
+
+                    REQUIRE(ta.get_tc().size() == 2u * (ta.get_order() + 1u) * batch_size);
+
+                    auto sa = xt::adapt(ta.get_state_data(), {2u, batch_size});
+                    auto isa = xt::adapt(init_state.data(), {2u, batch_size});
+
+                    {
+                        auto &oc = ta.step(true);
+
+                        for (auto i = 0u; i < batch_size; ++i) {
+                            auto ret = xt::eval(xt::zeros<double>({2u}));
+
+                            auto tca = xt::adapt(ta.get_tc(), {2u, ta.get_order() + 1u, batch_size});
+
+                            horner_eval(ret, xt::view(std::as_const(tca), xt::all(), xt::all(), i),
+                                        static_cast<int>(ta.get_order()), 0);
+                            REQUIRE(ret[0] == approximately(xt::view(isa, 0, i)[0], 10.));
+                            REQUIRE(ret[1] == approximately(xt::view(isa, 1, i)[0], 10.));
+
+                            horner_eval(ret, xt::view(std::as_const(tca), xt::all(), xt::all(), i),
+                                        static_cast<int>(ta.get_order()), std::get<1>(oc[i]));
+                            REQUIRE(ret[0] == approximately(xt::view(sa, 0, i)[0], 10.));
+                            REQUIRE(ret[1] == approximately(xt::view(sa, 1, i)[0], 10.));
+                        }
+                    }
+
+                    init_state = ta.get_state();
+
+                    {
+                        auto &oc = ta.step_backward(true);
+
+                        for (auto i = 0u; i < batch_size; ++i) {
+                            auto ret = xt::eval(xt::zeros<double>({2u}));
+
+                            auto tca = xt::adapt(ta.get_tc(), {2u, ta.get_order() + 1u, batch_size});
+
+                            horner_eval(ret, xt::view(std::as_const(tca), xt::all(), xt::all(), i),
+                                        static_cast<int>(ta.get_order()), 0);
+                            REQUIRE(ret[0] == approximately(xt::view(isa, 0, i)[0], 10.));
+                            REQUIRE(ret[1] == approximately(xt::view(isa, 1, i)[0], 10.));
+
+                            horner_eval(ret, xt::view(std::as_const(tca), xt::all(), xt::all(), i),
+                                        static_cast<int>(ta.get_order()), std::get<1>(oc[i]));
+                            REQUIRE(ret[0] == approximately(xt::view(sa, 0, i)[0], 10.));
+                            REQUIRE(ret[1] == approximately(xt::view(sa, 1, i)[0], 10.));
+                        }
+                    }
+
+                    init_state = ta.get_state();
+
+                    {
+                        std::vector<double> max_delta_t;
+                        for (auto i = 0u; i < batch_size; ++i) {
+                            max_delta_t.push_back(1e-5 + i * 1e-5);
+                        }
+
+                        auto &oc = ta.step(max_delta_t, true);
+
+                        for (auto i = 0u; i < batch_size; ++i) {
+                            auto ret = xt::eval(xt::zeros<double>({2u}));
+
+                            auto tca = xt::adapt(ta.get_tc(), {2u, ta.get_order() + 1u, batch_size});
+
+                            horner_eval(ret, xt::view(std::as_const(tca), xt::all(), xt::all(), i),
+                                        static_cast<int>(ta.get_order()), 0);
+                            REQUIRE(ret[0] == approximately(xt::view(isa, 0, i)[0], 10.));
+                            REQUIRE(ret[1] == approximately(xt::view(isa, 1, i)[0], 10.));
+
+                            horner_eval(ret, xt::view(std::as_const(tca), xt::all(), xt::all(), i),
+                                        static_cast<int>(ta.get_order()), std::get<1>(oc[i]));
+                            REQUIRE(ret[0] == approximately(xt::view(sa, 0, i)[0], 10.));
+                            REQUIRE(ret[1] == approximately(xt::view(sa, 1, i)[0], 10.));
+                        }
+                    }
                 }
             }
         }