+++ doc/changelog.rst
@@ -7,7 +7,7 @@ Changelog
 New
 ~~~
 
-- Implement support for continuous output
+- Implement support for dense output
   (`#92 <https://github.com/bluescarni/heyoka/pull/92>`__).
 - Add the ability to output the Taylor coefficients
   when invoking the single-step functions in the